Basic Platformer
I went off the tutorial for this
You can move with arrow keys and WASD, space to jump.
You can change the player speed and jump force in the unity scene, feel
free to play around with this as the physics aren't set in stone.
Right now the camera is a child of the player to follow it, but that can
easily be changed later