Level 1 Greybox Complete
I made some changes to the level while making it due to the types of
platforms I was given and lack of variety in the level.
The art assets are in, we still need something for the ground, water,
and floating logs or parts of a bridge.
Prefabs are set for all the sprites that have collisions so you can just
drag it onto the scene without any problems.